+++ src/components/NewsStand/Header/Header.jsx
@@ -1,37 +1,22 @@
 import React from 'react';
 import { Link, routes } from '@utils/Router';
-import { movingState, myNewsModeState, newsFetchState, newsPageSelector } from '@recoilStore/news';
-import { useRecoilState, useRecoilValue, useSetRecoilState } from 'recoil';
+import { myNewsModeState } from '@recoilStore/news';
+import { useSetRecoilState } from 'recoil';
 import { DirectionBtnWrap, HeaderWrap, MyNewsBtnWrap, NewsStandSortBtnWrap, RightControlBtnWrap } from './Header.style';
 import { HEADER_BTN, MODE } from '@utils/constant';
 import { useNewsMove, usePath, useTimer } from '@hooks';
+import { usePaging } from '../../../hooks/usePaging';
 
 const Header = () => {
   const { path } = usePath();
   const { moveNewsCompany } = useNewsMove();
+  const { movePaging } = usePaging();
   const setNewsMode = useSetRecoilState(myNewsModeState);
-  const [moving, setMoving] = useRecoilState(movingState);
-  const [page, setPage] = useRecoilState(newsPageSelector);
-  const lastPageIdx = Math.floor(useRecoilValue(newsFetchState).length / 15);
   useTimer({ path, moveNewsCompany });
+
   const handleClickArrow = ({ target: { name } }) => {
     if (path === routes.mynews.path) return moveNewsCompany(name);
-    if (path === routes.newscompany.path) {
-      if (moving) return;
-      setMoving(true);
-      if (name === 'prev') {
-        if (page === 0) {
-          return setPage(lastPageIdx);
-        }
-        setPage((prevPage) => prevPage - 1);
-      }
-      if (name === 'next') {
-        if (page === lastPageIdx) {
-          return setPage(0);
-        }
-        setPage((prevPage) => prevPage + 1);
-      }
-    }
+    if (path === routes.newscompany.path) return movePaging(name);
   };
 
   const changeViewMode = ({ target: { name } }) => setNewsMode(name);
